Impr Scraper class, selenium method, method refine
This commit has several new features. Methods which are functional have
been added to the class as functions. The functions now request
particular
variables depending on the function. In a previous commit, a list of
urls were generated with which to scrape multiple webpages. In this
commit, a method was defined to use Selenium to click through mulitple
webpages using an imputted button XPath.