Many web scraping functions added to scrape Class The function 'crypto_scrape_multi_days' to scrape data from coin market cap has been added to Class as a function. Within this function it is now possible to control how many records from each page to scrape, and this also be the maximum number by using the tr_length variable. This function also now packages each datum for each corresponding record into a single list, for example all Bitcoin data from a particular day, all Ethereum data from a particular day, etc..
There is now a function 'get_image_list_from_webpage' which gets src
links from a webpage and compiles them into a list. Another function,
'save_images_from_webpage' downloads and saves images from the scraped
src links.

There is now also a function 'UUID_dictionary' which generates a list of
unique UUIDs and takes an additional list (such as that from
'crypto_scrape_multi_days') in order to combine the records into a
dictionary. Each record then has a unique UUID associated with it. The
data is then saved locally

There is now a function 'list_zipper' which takes 2 lists and combines
them into a resulting list. This is useful, for example, to pair urls
of each scraped page to records, but could be broadly applicable to
any lists.